New PCI ID entries for 3c905 (0x9058) and DE-528. Fixed block number rollover bug in main.c due to signed promotion.

+++ src/main.c
@@ -741,7 +741,7 @@ int tftp(const char *name, int (*fnc)(unsigned char *, int, int, int))
 		else /* neither TFTP_OACK nor TFTP_DATA */
 			break;
 
-		if ((block || bcounter) && (block != prevblock+1)) {
+		if ((block || bcounter) && (block != (unsigned short)(prevblock+1))) {
 			/* Block order should be continuous */
 			tp.u.ack.block = htons(block = prevblock);
 		}
